feat(albumscontroller, alertinfo, albums.blade, alert-info.blade): add update resource message
In AlbumsController in update function save in $message the message of update and save $message in
session() and redirect to the albums.index. Then created component - php artisan make:component
alertInfo - that creates the component alertInfo.php and the component alert-info-blade.php. In
albums.blade  defined the component alert-info and pass the var message. In alertInfo class define
the variables and in alert-info.blade rendered the html.